makeValPlots: Script to compose plot-cmd for all RelVal plots.
There are a various number of RelVal plots which have to be created all
the time. Since there are all fairly similar, this script can be used to
create the plotting cmds with the right settings automatically.
Pipe them into a .sh file and execute it to get all relVal plots.
If Some of the plots look ugly, you can opn the .sh file and edit the
plot cmds manually.